refactor: add ui parts service to remove desktop ui controller
