+++ packages/ui/src/index.ts
@@ -28,7 +28,8 @@ export {
     UndoShortcutItem,
 } from './controllers/shared-shortcut.controller';
 export { IUIController, type IWorkbenchOptions } from './controllers/ui/ui.controller';
-export { DesktopUIController, DesktopUIPart, type IDesktopUIController } from './controllers/ui/ui-desktop.controller';
+export { DesktopUIController } from './controllers/ui/ui-desktop.controller';
+export { IUIPartsService, DesktopUIPart, DesktopUIPartsService } from './services/parts/parts.service';
 export { enUS, zhCN, ruRU } from './locale';
 export { DesktopBeforeCloseService, IBeforeCloseService } from './services/before-close/before-close.service';
 export { CopyCommand, CutCommand, PasteCommand } from './services/clipboard/clipboard.command';
