+++ packages/ui/src/ui-plugin.ts
@@ -55,6 +55,7 @@ import { IZenZoneService } from './services/zen-zone/zen-zone.service';
 import { EditorService, IEditorService } from './services/editor/editor.service';
 import { IRangeSelectorService, RangeSelectorService } from './services/range-selector/range-selector.service';
 import { IProgressService, ProgressService } from './services/progress/progress.service';
+import { DesktopUIPartsService, IUIPartsService } from './services/parts/parts.service';
 
 const PLUGIN_NAME = 'ui';
 
@@ -101,6 +102,7 @@ export class UniverUIPlugin extends Plugin {
 
             // services
             [ShortcutPanelService],
+            [IUIPartsService, { useClass: DesktopUIPartsService }],
             [ILayoutService, { useClass: DesktopLayoutService }],
             [IShortcutService, { useClass: DesktopShortcutService }],
             [IPlatformService, { useClass: DesktopPlatformService }],
